Revert "Move non-tree dependent tree-agent functions to new package" (#26862)
Reverts microsoft/FluidFramework#26862 and follow-up version fix
microsoft/FluidFramework#26914.
