+++ packages/framework/tree-agent/src/propertyBinding.ts
@@ -5,16 +5,18 @@
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import type { TreeNodeSchema } from "@fluidframework/tree";
-import {
-	exposePropertiesSymbol,
-	isTypeFactoryType,
-	type Ctor,
-	type ExposedProperties,
-	type IExposedProperties,
-	type TypeFactoryType,
-} from "@fluidframework/type-factory/internal";
 
-import type { BindableSchema } from "./methodBinding.js";
+import type { BindableSchema, Ctor } from "./methodBinding.js";
+import type { TypeFactoryType } from "./treeAgentTypes.js";
+import { isTypeFactoryType } from "./treeAgentTypes.js";
+
+/**
+ * A symbol used to expose properties to the LLM.
+ * @alpha
+ */
+export const exposePropertiesSymbol: unique symbol = Symbol.for(
+	"@fluidframework/tree-agent/exposeProperties",
+);
 
 /**
  * A property definition class that describes the structure of the property
@@ -41,20 +43,65 @@ export class PropertyDef {
 	) {}
 }
 
+/**
+ * An interface for exposing properties of schema classes to an agent.
+ * @alpha
+ */
+export interface ExposedProperties {
+	/**
+	 * Expose a property with type factory type and metadata.
+	 */
+	exposeProperty<S extends BindableSchema & Ctor, K extends string>(
+		schema: S,
+		name: K,
+		def: { schema: TypeFactoryType; description?: string; readOnly?: boolean },
+	): void;
+
+	/**
+	 * Expose a property with type factory type (simple form).
+	 */
+	exposeProperty<S extends BindableSchema & Ctor, K extends string>(
+		schema: S,
+		name: K,
+		tfType: TypeFactoryType,
+	): void;
+}
+
+/**
+ * An interface that SharedTree schema classes should implement to expose their properties to the LLM.
+ *
+ * @remarks
+ * The `getExposedProperties` free function will cause the method here to be called on the class passed to it.
+ *
+ * @privateremarks
+ * Implementing this interface correctly seems tricky?
+ * To actually implement it in a way that satisfies TypeScript,
+ * classes need to declare both a static version and an instance version of the method
+ * (the instance one can just delegate to the static one).
+ *
+ * @alpha
+ */
+export interface IExposedProperties {
+	/**
+	 * Static method that exposes properties of this schema class to an agent.
+	 */
+	[exposePropertiesSymbol]?(properties: ExposedProperties): void;
+}
+
 class ExposedPropertiesI implements ExposedProperties {
 	private readonly properties: Record<string, PropertyDef> = {};
 	private readonly referencedTypes = new Set<TreeNodeSchema>();
 
 	public constructor(private readonly schemaClass: BindableSchema) {}
 
-	public exposeProperty<S extends Ctor, K extends string>(
+	public exposeProperty<S extends BindableSchema & Ctor, K extends string>(
 		schema: S,
 		name: K,
 		defOrType:
 			| { schema: TypeFactoryType; description?: string; readOnly?: boolean }
 			| TypeFactoryType,
 	): void {
-		if ((schema as unknown) !== (this.schemaClass as unknown)) {
+		if (schema !== this.schemaClass) {
 			throw new Error('Must expose properties on the "this" schema class');
 		}
