+++ packages/framework/tree-agent/src/renderSchemaTypeScript.ts
@@ -13,18 +13,14 @@ import {
 } from "@fluidframework/tree/alpha";
 import { FieldKind, NodeKind, ValueSchema } from "@fluidframework/tree/internal";
 import type { SimpleLeafNodeSchema } from "@fluidframework/tree/internal";
-import {
-	isTypeFactoryType,
-	type TypeFactoryOptional,
-	type TypeFactoryType,
-} from "@fluidframework/type-factory/internal";
 
 import type { BindableSchema, FunctionWrapper } from "./methodBinding.js";
 import { getExposedMethods } from "./methodBinding.js";
 import { fluidHandleTypeName } from "./prompt.js";
-import type { PropertyDef } from "./propertyBinding.js";
-import { getExposedProperties } from "./propertyBinding.js";
+import { getExposedProperties, type PropertyDef } from "./propertyBinding.js";
 import { renderTypeFactoryTypeScript } from "./renderTypeFactoryTypeScript.js";
+import type { TypeFactoryOptional, TypeFactoryType } from "./treeAgentTypes.js";
+import { isTypeFactoryType } from "./treeAgentTypes.js";
 import {
 	IdentifierCollisionResolver,
 	getFriendlyName,
