+++ packages/framework/tree-agent/src/test/typeFactoryRendering.spec.ts
@@ -6,10 +6,9 @@
 import { strict as assert } from "node:assert";
 
 import { SchemaFactory } from "@fluidframework/tree/internal";
-import type { TypeFactoryType } from "@fluidframework/type-factory/alpha";
 
 import { renderTypeFactoryTypeScript } from "../renderTypeFactoryTypeScript.js";
-import { typeFactory as tf } from "../treeAgentTypes.js";
+import { typeFactory as tf, type TypeFactoryType } from "../treeAgentTypes.js";
 
 const sf = new SchemaFactory("test");
