+++ packages/framework/tree-agent/src/test/typeGeneration.spec.ts
@@ -13,15 +13,10 @@ import {
 	type ImplicitFieldSchema,
 	type InsertableField,
 } from "@fluidframework/tree/internal";
-import {
-	buildFunc,
-	exposeMethodsSymbol,
-	exposePropertiesSymbol,
-	type ExposedMethods,
-	type ExposedProperties,
-} from "@fluidframework/type-factory/alpha";
 
+import { buildFunc, exposeMethodsSymbol, type ExposedMethods } from "../methodBinding.js";
 import { fluidHandleTypeName } from "../prompt.js";
+import { exposePropertiesSymbol, type ExposedProperties } from "../propertyBinding.js";
 import { typeFactory } from "../treeAgentTypes.js";
 import { generateEditTypesForPrompt } from "../typeGeneration.js";
 
@@ -36,7 +31,7 @@ class Todo extends sf.object("Todo", {
 	}
 
 	public static [exposeMethodsSymbol](methods: ExposedMethods): void {
-		methods.exposeMethod(
+		methods.expose(
 			Todo,
 			"method",
 			buildFunc({ returns: typeFactory.boolean() }, ["n", typeFactory.string()]),
@@ -50,7 +45,7 @@ class TestTodoAppSchema extends sf.object("TestTodoAppSchema", {
 	todos: sf.array(Todo),
 }) {
 	public static [exposeMethodsSymbol](methods: ExposedMethods): void {
-		methods.exposeMethod(
+		methods.expose(
 			TestTodoAppSchema,
 			"addTodo",
 			buildFunc({ returns: typeFactory.instanceOf(Todo) }, [
@@ -104,7 +99,7 @@ describe("Type generation", () => {
 				}
 
 				public static [exposeMethodsSymbol](methods: ExposedMethods): void {
-					methods.exposeMethod(
+					methods.expose(
 						ObjWithMethod,
 						"method",
 						buildFunc({ returns: typeFactory.boolean() }, ["n", typeFactory.string()]),
@@ -129,7 +124,7 @@ describe("Type generation", () => {
 				}
 
 				public static [exposeMethodsSymbol](methods: ExposedMethods): void {
-					methods.exposeMethod(
+					methods.expose(
 						ArrayWithMethod,
 						"method",
 						buildFunc({ returns: typeFactory.boolean() }, ["n", typeFactory.string()]),
@@ -155,7 +150,7 @@ type ArrayWithMethod = string[] & {
 				}
 
 				public static [exposeMethodsSymbol](methods: ExposedMethods): void {
-					methods.exposeMethod(
+					methods.expose(
 						MapWithMethod,
 						"method",
 						buildFunc({ returns: typeFactory.boolean() }, ["n", typeFactory.string()]),
@@ -183,7 +178,7 @@ type MapWithMethod = Map<string, string> & {
 				}
 
 				public static [exposeMethodsSymbol](methods: ExposedMethods): void {
-					methods.exposeMethod(
+					methods.expose(
 						MapWithMethod,
 						"method",
 						buildFunc({ returns: typeFactory.instanceOf(Obj) }, ["n", typeFactory.string()]),
