--- packages/framework/type-factory/.mocharc.cjs
@@ -1,11 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
- * Licensed under the MIT License.
- */
-
-"use strict";
-
-const getFluidTestMochaConfig = require("@fluid-internal/mocha-test-setup/mocharc-common");
-
-const config = getFluidTestMochaConfig(__dirname);
-module.exports = config;
