--- packages/framework/type-factory/.npmignore
@@ -1,7 +1,0 @@
-nyc
-*.log
-**/*.tsbuildinfo
-src/test
-dist/test
-lib/test
-**/_api-extractor-temp/**
