--- packages/framework/type-factory/api-extractor.json
@@ -1,14 +1,0 @@
-{
-	"$schema": "https://developer.microsoft.com/json-schemas/api-extractor/v7/api-extractor.schema.json",
-	"extends": "../../../common/build/build-common/api-extractor-base.esm.no-legacy.json",
-	"compiler": {
-		// api-extractor has trouble with Node16 moduleResolution, but Bundler looks okay.
-		"overrideTsconfig": {
-			"$schema": "http://json.schemastore.org/tsconfig",
-			"extends": "./tsconfig.json",
-			"compilerOptions": {
-				"moduleResolution": "Bundler"
-			}
-		}
-	}
-}
