--- packages/framework/type-factory/api-report/type-factory.alpha.api.md
@@ -1,204 +1,0 @@
-## Alpha API Report File for "@fluidframework/type-factory"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-
-// @alpha
-export type Arg<T extends TypeFactoryType = TypeFactoryType> = readonly [
-name: string,
-type: T
-];
-
-// @alpha
-export function buildFunc<const Return extends TypeFactoryType, const Args extends readonly Arg[], const Rest extends TypeFactoryType | null = null>(def: {
-    description?: string;
-    returns: Return;
-    rest?: Rest;
-}, ...args: Args): FunctionDef<Args, Return, Rest>;
-
-// @alpha
-export type Ctor<T = any> = new (...args: any[]) => T;
-
-// @alpha
-export interface ExposedMethods {
-    exposeMethod<const K extends string & keyof MethodKeys<InstanceType<S>>, S extends Ctor & IExposedMethods, Z extends FunctionDef<readonly Arg[], TypeFactoryType, TypeFactoryType | null>>(constructor: S, methodName: K, tfFunction: Z): void;
-}
-
-// @alpha
-export interface ExposedProperties {
-    exposeProperty<S extends Ctor, K extends string>(constructor: S, name: K, def: {
-        schema: TypeFactoryType;
-        description?: string;
-        readOnly?: boolean;
-    }): void;
-    exposeProperty<S extends Ctor, K extends string>(constructor: S, name: K, tfType: TypeFactoryType): void;
-}
-
-// @alpha
-export const exposeMethodsSymbol: unique symbol;
-
-// @alpha
-export const exposePropertiesSymbol: unique symbol;
-
-// @alpha
-export interface FunctionDef<Args extends readonly Arg[], Return extends TypeFactoryType, Rest extends TypeFactoryType | null = null> {
-    args: Args;
-    description?: string;
-    rest?: Rest;
-    returns: Return;
-}
-
-// @alpha
-export interface IExposedMethods {
-    [exposeMethodsSymbol](methods: ExposedMethods): void;
-}
-
-// @alpha
-export interface IExposedProperties {
-    [exposePropertiesSymbol]?(properties: ExposedProperties): void;
-}
-
-// @alpha
-export function isTypeFactoryType(value: unknown): value is TypeFactoryType;
-
-// @alpha
-export type MethodKeys<T> = {
-    [K in keyof T]: T[K] extends (...args: any[]) => any ? K : never;
-};
-
-// @alpha
-export interface TypeFactoryArray extends TypeFactoryType {
-    readonly element: TypeFactoryType;
-    readonly _kind: "array";
-}
-
-// @alpha
-export interface TypeFactoryBoolean extends TypeFactoryType {
-    readonly _kind: "boolean";
-}
-
-// @alpha
-export interface TypeFactoryDate extends TypeFactoryType {
-    readonly _kind: "date";
-}
-
-// @alpha
-export interface TypeFactoryFunction extends TypeFactoryType {
-    readonly _kind: "function";
-    readonly parameters: readonly TypeFactoryFunctionParameter[];
-    readonly restParameter?: TypeFactoryFunctionParameter;
-    readonly returnType: TypeFactoryType;
-}
-
-// @alpha
-export type TypeFactoryFunctionParameter = readonly [name: string, type: TypeFactoryType];
-
-// @alpha
-export interface TypeFactoryInstanceOf extends TypeFactoryType {
-    readonly constructor: new (...args: any[]) => unknown;
-    readonly _kind: "instanceof";
-}
-
-// @alpha
-export interface TypeFactoryIntersection extends TypeFactoryType {
-    readonly _kind: "intersection";
-    readonly types: readonly TypeFactoryType[];
-}
-
-// @alpha
-export interface TypeFactoryLiteral extends TypeFactoryType {
-    readonly _kind: "literal";
-    readonly value: string | number | boolean;
-}
-
-// @alpha
-export interface TypeFactoryMap extends TypeFactoryType {
-    readonly keyType: TypeFactoryType;
-    readonly _kind: "map";
-    readonly valueType: TypeFactoryType;
-}
-
-// @alpha
-export interface TypeFactoryNull extends TypeFactoryType {
-    readonly _kind: "null";
-}
-
-// @alpha
-export interface TypeFactoryNumber extends TypeFactoryType {
-    readonly _kind: "number";
-}
-
-// @alpha
-export interface TypeFactoryObject extends TypeFactoryType {
-    readonly _kind: "object";
-    readonly shape: Record<string, TypeFactoryType>;
-}
-
-// @alpha
-export interface TypeFactoryOptional extends TypeFactoryType {
-    readonly innerType: TypeFactoryType;
-    readonly _kind: "optional";
-}
-
-// @alpha
-export interface TypeFactoryPromise extends TypeFactoryType {
-    readonly innerType: TypeFactoryType;
-    readonly _kind: "promise";
-}
-
-// @alpha
-export interface TypeFactoryReadonly extends TypeFactoryType {
-    readonly innerType: TypeFactoryType;
-    readonly _kind: "readonly";
-}
-
-// @alpha
-export interface TypeFactoryRecord extends TypeFactoryType {
-    readonly keyType: TypeFactoryType;
-    readonly _kind: "record";
-    readonly valueType: TypeFactoryType;
-}
-
-// @alpha
-export interface TypeFactoryString extends TypeFactoryType {
-    readonly _kind: "string";
-}
-
-// @alpha
-export interface TypeFactoryTuple extends TypeFactoryType {
-    readonly items: readonly TypeFactoryType[];
-    readonly _kind: "tuple";
-    readonly rest?: TypeFactoryType;
-}
-
-// @alpha
-export interface TypeFactoryType {
-    readonly _kind: TypeFactoryTypeKind;
-}
-
-// @alpha
-export type TypeFactoryTypeKind = "string" | "number" | "boolean" | "void" | "undefined" | "null" | "unknown" | "date" | "promise" | "array" | "object" | "record" | "map" | "tuple" | "union" | "intersection" | "literal" | "optional" | "readonly" | "function" | "instanceof";
-
-// @alpha
-export interface TypeFactoryUndefined extends TypeFactoryType {
-    readonly _kind: "undefined";
-}
-
-// @alpha
-export interface TypeFactoryUnion extends TypeFactoryType {
-    readonly _kind: "union";
-    readonly options: readonly TypeFactoryType[];
-}
-
-// @alpha
-export interface TypeFactoryUnknown extends TypeFactoryType {
-    readonly _kind: "unknown";
-}
-
-// @alpha
-export interface TypeFactoryVoid extends TypeFactoryType {
-    readonly _kind: "void";
-}
-
-```
