--- packages/framework/type-factory/api-report/type-factory.beta.api.md
@@ -1,7 +1,0 @@
-## Beta API Report File for "@fluidframework/type-factory"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-
-```
