--- packages/framework/type-factory/eslint.config.mts
@@ -1,43 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
- * Licensed under the MIT License.
- */
-
-import type { Linter } from "eslint";
-import { strict } from "../../../common/build/eslint-config-fluid/flat.mts";
-
-const config: Linter.Config[] = [
-	...strict,
-	{
-		rules: {
-			"import-x/no-internal-modules": [
-				"error",
-				{
-					"allow": [
-						"@fluidframework/*/alpha",
-						"@fluidframework/*/beta",
-						"@fluidframework/*/internal",
-					],
-				},
-			],
-		},
-	},
-	{
-		files: ["src/test/**/*"],
-		rules: {
-			"import-x/no-internal-modules": [
-				"error",
-				{
-					"allow": [
-						"*/index.js",
-						"@fluidframework/*/alpha",
-						"@fluidframework/*/beta",
-						"@fluidframework/*/internal",
-					],
-				},
-			],
-		},
-	},
-];
-
-export default config;
