--- packages/framework/type-factory/src/test/tsconfig.cjs.json
@@ -1,12 +1,0 @@
-{
-	// This config must be used in a "type": "commonjs" environment. (Use `fluid-tsc commonjs`.)
-	"extends": "./tsconfig.json",
-	"compilerOptions": {
-		"outDir": "../../dist/test",
-	},
-	"references": [
-		{
-			"path": "../../tsconfig.cjs.json",
-		},
-	],
-}
