--- packages/framework/type-factory/src/test/tsconfig.json
@@ -1,16 +1,0 @@
-{
-	"extends": "../../../../../common/build/build-common/tsconfig.test.node16.json",
-	"compilerOptions": {
-		"rootDir": "./",
-		"outDir": "../../lib/test",
-		"types": ["node", "mocha"],
-		"declaration": true,
-		"declarationMap": true,
-	},
-	"include": ["./**/*"],
-	"references": [
-		{
-			"path": "../../tsconfig.json",
-		},
-	],
-}
