--- packages/framework/type-factory/tsconfig.cjs.json
@@ -1,7 +1,0 @@
-{
-	// This config must be used in a "type": "commonjs" environment. (Use `fluid-tsc commonjs`.)
-	"extends": "./tsconfig.json",
-	"compilerOptions": {
-		"outDir": "./dist",
-	},
-}
