--- packages/framework/type-factory/tsconfig.json
@@ -1,14 +1,0 @@
-{
-	"extends": "../../../common/build/build-common/tsconfig.node16.json",
-	"include": ["src/**/*"],
-	"exclude": ["src/test/**/*"],
-	"compilerOptions": {
-		"rootDir": "./src",
-		"outDir": "./lib",
-		"noImplicitAny": true,
-		"preserveConstEnums": true,
-		"exactOptionalPropertyTypes": false,
-		"lib": ["ES2020", "DOM", "DOM.Iterable"],
-		"skipLibCheck": true,
-	},
-}
