--- packages/framework/type-factory/tsdoc.json
@@ -1,4 +1,0 @@
-{
-	"$schema": "https://developer.microsoft.com/json-schemas/tsdoc/v0/tsdoc.schema.json",
-	"extends": ["../../../common/build/build-common/tsdoc-base.json"]
-}
